Workaround for race condition, also restored `write` support.
I found that a race condition can occur when the writer does a lot
of processing. I'm not entirely certain how and why it occurs.
It's difficult to debug because it's in threaded code, and also
because stdout and stderr end up being unavailable because of
the nature of what streamcapture does. Nevertheless, using webpdb,
I was able to get some information.

When the parent thread calls StreamCapture.close() or
FDCapture.close(), this immediately closes the write
end of the pipe. In principle, the thread that reads
from that pipe should detect and EOF and terminate.
Only when all the slave threads have thus terminated, can the
overall process exit. I found that the slave threads were not
getting an EOF and the overall process was thus stuck. It was
even quite hard to kill it with ctrl-c or such.

Debugging some more, I found that when the parent thread
calls os.close() on the write end of the pipe, sometimes
the pipe does not get closed and yet os.close() does not
signal an error. The underlying C library close() function
is indeed allowed to error out instead of closing an fd,
but I did not get to the bottom of this to see whether
indeed this is happening. One obstacle is that python
does not expose `errno` directly.

From the debugger, if I try to call os.close() again on the
write end of the pipe, it usually then works. So I have
no idea why it doesn't work the first time around.

I also tried to call os.close() until the underlying fd
is reported as closed by fstat. This did indeed close an fd
but then the whole python interpreter core dumped somewhere
else. So there might be some weird underlying python bug,
or even an OS bug.

As a workaround, I've decided to send a binary message
through the pipe, that should not normally occur in a
UTF-8 stream. This magic message is b'\x04\x00\x08\x20',
and when the pipe reading thread receives it, it proceeds
as if the pipe had been closed and terminates.

Also, I've re-added support for writer.write() so that
a plain-old stream can be passed to StreamCapture. You
don't need to wrap those inside of a Writer if you
don't want to.