Updated magic EOF string
The magic EOF string has been updated to:
b'\x04\x81\x00\xff'

Not only should this string not appear in UTF-8 sequences, it has
a bunch of illegal bytes in it. Inspiration was taken from here:
https://www.cl.cam.ac.uk/~mgk25/ucs/examples/UTF-8-test.txt

The motivation:
\x04 is ctrl-D, which is the Unix EOF character.
\x81 is a UTF-8 continuation byte, but it is not part of a
multibyte UTF-8 symbol so it is illegal and should not occur.
Furthermore, according to the ASCII web page below, \x81 is
unused in more basic ASCII stuff.
https://www.ascii-code.com/
\x00 is unlikely to be emitted by C programs because of
null termination of strings.
\xff is also an illegal UTF-8 character.

Of course, it is possible to send such byte sequences to stdout
in a legitimate fashion (e.g. by cat-ing a binary file) but until
I can figure out whatever race condition/python bug/OS bug,
I can't rely on os.close() to close the write end of pipes.